Implement exitless IPI
The exitless IPI is implemented by the posted-interrupt
mechanism. Typically, the guest needs the host help to send
IPI. When the guest writes to the APIC ICR, it triggers the VM
exit. KVM uses the APIC destination ID saved in the VMCS and
sends the IPI for the guest. If there is posted-interrupt
support, such IPI is delivered as the posted interrupt to the
destination vCPU. It does not preempt the running vCPU at the
destination.

How It Works
------------
- KVM exposes the hardware APIC ID to the vAPICs of each vCPU.
- KVM updates the emulated IO-APIC for the emulated
  interrupts.
- Guest updates per-vcpu x86_cpu_to_apicid.
- Guest enables the DTID.
  * function pointer redirection.
  * disable wrmsr tmict vm exit.
- Guest enables the direct IPI.
  * function pointer redirection.
  * disable wrmsr icr vm exit.

modified:   did.c
modified:   did.h
new file:   did.sh
modified:   run_did.c