Use test_and_set_bit for DID
test_and_set_bit
----------------
This operation is atomic and cannot be reordered. It also
implies a memory barrier. Previously, this reduced the network
bandwidth of Infiniband. We hypothesized that the cache-line
lock interferes how KVM and VT-d chipset operated on the
shared PID page. For a mysterious and unknown reason, it works
fine now.

Data: outbound (Gbps)/inbound(Gbps)/rtt(us)
- host: 37.6/37.6/12
- vanilla guest: 37.6/37.6/20
- dtid guest: 37.6/37.6/15
- did guest: 37.6/37.6/14

__set_bit
---------
This function is non-atomic and may be reordered. If it's
called on the same region of memory simultaneously, the effect
may be that only one operation succeeds.

modified:   did.c